Turtle Module Usage
This Python project utilizes the Turtle module to create colorful geometric shapes with varying numbers of sides. Each shape is drawn using the Turtle graphics library, and the number of sides ranges from 3 to 10.

The program randomly selects colors from a predefined list and draws polygons with a specified number of sides. The main features of the project include:

- Drawing polygons with sides ranging from 3 to 10.
- Random selection of colors from a predefined list.
- Utilization of the Turtle graphics module for drawing.